Correção dos apontamentos do PR B2WM 1451
