+++ YDMOfflineAccount/Home/View/HomeHeaderViewCell.swift
@@ -11,11 +11,30 @@ import YDB2WModels
 
 class HomeHeaderViewCell: UICollectionReusableView {
   
+  // MARK: Views
   var userProfileView = YDUserProfileView()
   
-  func initView(user: YDCurrentCustomer) {
+  // MARK: Init
+  override init(frame: CGRect) {
+    super.init(frame: .zero)
+    configUI()
+  }
+  
+  required init?(coder: NSCoder) {
+    fatalError("init(coder:) has not been implemented")
+  }
+  
+  // MARK: Functions
+  func populateView(user: YDCurrentCustomer) {
     userProfileView.config(username: user.fullName ?? "", userPhoto: nil)
-    
+  }
+  
+}
+
+// MARK: UI
+extension HomeHeaderViewCell {
+  
+  private func configUI() {
     userProfileView.translatesAutoresizingMaskIntoConstraints = false
     self.addSubview(userProfileView)
     NSLayoutConstraint.activate([
@@ -24,7 +43,6 @@ class HomeHeaderViewCell: UICollectionReusableView {
       userProfileView.topAnchor.constraint(equalTo: self.topAnchor, constant: 32),
       userProfileView.bottomAnchor.constraint(equalTo: self.bottomAnchor, constant: -24)
     ])
-    
   }
   
 }
