+++ YDMOfflineAccount/Home/ViewController/Extensions/CollectionView/HomeViewController+CollectionView.swift
@@ -76,7 +76,7 @@ extension HomeViewController: UICollectionViewDelegate, UICollectionViewDataSour
       for: indexPath
     ) as! HomeHeaderViewCell
     
-    header.initView(user: viewModel!.currentUser)
+    header.populateView(user: viewModel!.currentUser)
     
     return header
